pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hoffmagic"
version = "0.1.0"
description = "A beautiful Python-based blog application"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.2",
    "jinja2>=3.1.2",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "psycopg>=3.1.12",
    "python-multipart>=0.0.6",
    "markdown>=3.5.1",
    "pygments>=2.16.1",
    "pillow>=10.1.0",
    "python-frontmatter>=1.0.0",
    "email-validator>=2.1.0",
    "typer>=0.9.0",
    "rich>=13.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.10.1",
    "isort>=5.12.0",
    "ruff>=0.1.3",
    "mypy>=1.6.1",
]

[project.scripts]
hoffmagic = "hoffmagic.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I", "N", "UP", "ANN", "RUF"]
ignore = ["ANN101", "ANN401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
